% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cell_x_adt}
\alias{cell_x_adt}
\title{A matrix of raw count for the cell by ADT markers}
\format{
A data frame with 422682 rows and 9 variables:
\describe{
\item{CD3}{CD3 ADT marker raw count across each cell}
\item{CD4}{CD4 ADT marker raw count across each cell}
\item{CD8}{CD8 ADT marker raw count across each cell}
\item{CD14}{CD14 ADT marker raw count across each cell}
\item{CD19}{CD19 ADT marker raw count across each cell}
\item{CD25}{CD25 ADT marker raw count across each cell}
\item{CD45RA}{CD45RA ADT marker raw count across each cell}
\item{CD56}{CD56 ADT marker raw count across each cell}
\item{CD127}{CD127 ADT marker raw count across each cell}
}
}
\source{
See detailed description in the manuscript
}
\usage{
cell_x_adt
}
\description{
A dataset containing 422682 cells and 9 ADT markers for the CITE-seq raw measurement of 13 public available CITE-seq datasets.
}
\keyword{datasets}
