% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_adt_name.R
\name{clean_adt_name}
\alias{clean_adt_name}
\title{Clean ADT marker name}
\usage{
clean_adt_name(adt_name)
}
\arguments{
\item{adt_name}{A vector of ADT marker name.}
}
\description{
This function enables the general ADT name clearning but users should try cleaning and unifying their ADT marker name first. Please also make sure there is no "/" in the ADT name such as "TCRγ/δ".
}
\examples{
\dontrun{
clean_adt_name(colnames(cell_x_adt))
}
}
