% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_impute_outlier_valley.R
\name{detect_impute_outlier_valley}
\alias{detect_impute_outlier_valley}
\title{Identify the valley outliers and impute by valley by closet neighbor samples on the graph.}
\usage{
detect_impute_outlier_valley(
  valley_location_res,
  cell_x_feature,
  scale = 3,
  method = "MAD",
  nearest_neighbor_n = 3,
  nearest_neighbor_threshold = 0.75
)
}
\arguments{
\item{valley_location_res}{Matrix of valley landmark locations with rows being samples and columns being the valleys.}

\item{cell_x_feature}{Matrix of cells (rows) by cell features (columns) such as cell type, sample, and batch related information.}

\item{scale}{Scale level to defining outlier. Larger scale value corresponds to more severe ourliers.}

\item{method}{Outlier detection methods, choose from "MAD" (Median Absolute Deviation) or "IQR" (InterQuartile Range).}

\item{nearest_neighbor_n}{Number of top nearest neighbor samples to detect.}

\item{nearest_neighbor_threshold}{Threshold to call neighbor samples.}
}
\description{
This function identify the valley that tend to be outliers compared to other valley locations and try to find the closest samples that have similar density distribution to impute the valley. If no neighbor sample is detected, the valley will remain as original.
}
\examples{
detect_impute_outlier_valley(valley_location_res, cell_x_feature)
}
