% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_customize_landmark.R
\name{flat_table}
\alias{flat_table}
\title{Get the user-customized peaks and valleys location for each sample}
\usage{
flat_table(tab)
}
\arguments{
\item{cell_x_adt_sample}{Matrix of adt counts of seleted marker, with columns of sample and batch information for each row of cells.}

\item{landmark_pos}{Matrix of landmark location including peaks and valleys.}

\item{bw}{Bandwidth for the density plot.}

\item{adt_marker_select_name}{The adt marker needed to be manually processed to set the landmarks.}

\item{brewer_palettes}{Set the color scheme of color brewer. Default is "Set1".}
}
\description{
This function will launch a shiny app to allow user to manually set the peaks and valleys location. The function will output the landmark positions that user has set.
}
\examples{
\dontrun{
get_peak_mode(
  cell_x_adt = cell_x_adt,
  cell_x_feature = cell_x_feature,
  adt_marker_select = "CD3",
  neg_candidate_thres = asinh(6/5+1)
)
}
}
