% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_landmarks.R
\name{load_landmarks}
\alias{load_landmarks}
\title{Load saved landmarks into a list of matrices, as required for input into "override_landmark".}
\usage{
load_landmarks(dir, append_rds = TRUE)
}
\arguments{
\item{dir}{Directory where landmark location .rds files are stored}

\item{append_rds}{Whether to append "/RDS" to the provided directory path. The default is TRUE, which aligns with ADTnorm's default save location.}
}
\description{
Load saved landmarks into a list of matrices, as required for input into "override_landmark".
}
\examples{
\dontrun{
override_landmark = load_landmarks(dir = save_outpath)
}
}
