% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_adt_density_each.R
\name{plot_adt_density_each}
\alias{plot_adt_density_each}
\title{Plot the expression density profile for ONE ADT marker}
\usage{
plot_adt_density_each(
  adt_count,
  cell_x_feature,
  brewer_palettes,
  parameter_list = NULL
)
}
\arguments{
\item{adt_count}{Matrix of ADT raw counts in cells (rows) by one target ADT marker (column) format.}

\item{cell_x_feature}{Matrix of cells (rows) by cell features (columns) such as cell type, sample, and batch related information.}

\item{brewer_palettes}{Set the color scheme of color brewer.}

\item{parameter_list}{Users can specify: "run_label" to give name for this run; "bw" to adjust the band width of the density plot.}
}
\description{
This function plots adt expression density profile for only one ADT marker. Each track is a sample. Color by batch
}
\examples{
\dontrun{
plot_adt_density_each(
 cell_x_adt,
 cell_x_feature,
 brewer_palettes = "Set1",
 parameter_list = list(bw = 0.1, run_label = "ADTnorm")
)
}
}
