% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_alignment.R
\name{peak_alignment}
\alias{peak_alignment}
\title{Align the peak and valley landmarks by the warpset function}
\usage{
peak_alignment(
  cell_x_adt,
  cell_x_feature = NULL,
  landmark_matrix = NULL,
  target_landmark = NULL
)
}
\arguments{
\item{cell_x_adt}{Matrix of ADT raw counts in cells (rows) by ADT markers (columns) format.}

\item{cell_x_feature}{Matrix of cells (rows) by cell features (columns) such as cell type, sample, and batch related information.}

\item{landmark_matrix}{Matrix of peak and valley landmarks after fill in NA using \code{landmark_fill_na} function.}

\item{target_landmark}{Leave it as NULL to align the landmark to the mean location across samples. Denote it by a vector of the same length of the column number of landmark to align the negative peak, valley and positive peak(s) to the specified fixed location.}
}
\description{
This function detect the valley locations either between every two peak landmarks or cut at the right heavy tails. If specified positive uni-peak, the valley location will be set at the left side of the uni-peak.
}
\examples{
\dontrun{
peak_alignment(cell_x_adt, cell_x_feature, landmark_matrix)
}
}
